refactor(room manager): extract `RoomManager` from `Room.js`
`Room.js`파일이 너무 커져서, `RoomManager`를 분리함.
